requirements.txt
loguru
requests
undetected_chromedriver
python-dotenv